% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDailyDosage.R, R/funPharm.R
\name{calDailyDosage}
\alias{calDailyDosage}
\title{Calculate DDDs in RxCui}
\usage{
calDailyDosage(
  df,
  RxCuiColName = RxCui,
  QuantityColName = Quantity,
  DaysSupplyConName = DaysSupply,
  cores = 4
)
}
\arguments{
\item{df}{data.frame, a drug ingredient`s ATC of the specific indication}

\item{RxCuiColName}{a column name for RxCui of df}

\item{QuantityColName}{a column name for drug quantity of df}

\item{DaysSupplyConName}{a column name for the days that patient`s drug supplied of df}
}
\value{
A new data.table \code{sample_data_subset_atc} with following columns.

\code{MemberId} patient Id

\code{DispenseDate} patient`s drug dispensed date, and the time interval is from 2025-12-06 to 2038-01-25

\code{NationalDrugCode} drug NDC, total of 17,326 different drug codes

\code{Quantity} quantity of patient`s drug dispensed

\code{DaysSupply} days of patient`s drug supplied

\code{RxCui} drug RxCui

\code{ATC} drug ATC

\code{DailyDosage} daily dosage drugs

\code{Unit} drug calculate unit
}
\description{
\code{calDailyDosage} is a function to calculate daily dosage drugs(DDDs) in RxCui
}
\details{
This function provides user to simply calculate daily dosage for drug in RxCui.

The calculation model: The dosage of the drug multiplied by the number of the days of drug taken, and then divided by the number of days of drug supplied.
}
\examples{
# sample of calculating DDDs in RxCui.
sample_data_subset_atc <- calDailyDosage(df = sample_data_subset_atc, RxCuiColName = RxCui, QuantityColName = Quantity, DaysSupplyConName = DaysSupply, cores = 2)
head(sample_data_subset_atc)
}
