% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDDDs_range.R
\name{calDDDs.range}
\alias{calDDDs.range}
\title{Calculate subjects' ddd in a period}
\usage{
calDDDs.range(case, index_dayColName = index_day,
  expo_range_before = 36000, expo_range_after = 36000,
  idColName = patient_id, DispensingColName = Dispensing,
  SupplyDayColName = duration, DailyDosageColName = Daily_dosage)
}
\arguments{
\item{case}{data.frame include subjects' id, dispensing date, drug ATC code, daily dosage, duration}

\item{expo_range_before}{days before observation day}

\item{expo_range_after}{days after observation day}

\item{idColName}{a colum for subject's id}

\item{DispensingColName}{a colum for dispensing}

\item{SupplyDayColName}{a colum for supply day}

\item{DailyDosageColName}{a colum for daily dosage}

\item{index_day}{observation day}
}
\description{
Calculate subjects' ddd in a period
}
