% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDDDs.R
\name{calDDDs}
\alias{calDDDs}
\title{Calculate subjects' ddd}
\usage{
calDDDs(case)
}
\arguments{
\item{case}{data.frame include subjects' id, dispensing date, drug ATC code, daily dosage, duration}
}
\description{
Calculate subjects' ddd
}
