% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDDDsAccumulation.R, R/funPharm.R,
%   R/getDDDs.R
\name{calDDDsAccumulation}
\alias{calDDDsAccumulation}
\alias{getDDDs}
\title{Calculate subjects' accumulated DDDs}
\usage{
calDDDsAccumulation(
  case,
  PatientIdColName = MemberId,
  DispensingColName = DispenseDate,
  AtcCodeColName = ATC,
  DailyDosageColName = DailyDosage,
  DurationColName = DaysSupply
)

getDDDs(atc)
}
\arguments{
\item{case}{data.frame, include subjects' id, dispensing date, drug ATC code, daily dosage, duration}

\item{atc}{data.frame include ATC code}
}
\value{
A new data.table \code{accumulat_DDD} with following columns.

\code{MemberId} patient Id

\code{DDDs} accumulated DDDs
}
\description{
\code{calDailyDosage} is a function to Calculate subjects' accumulated daily dosage drugs(DDDs) in RxCui before last dispensed
\code{getDDDs} is a function to get DDDs in ATC
}
\details{
This function provides user to calculate patient`s total amount of DDDs for accumulation medication time.
}
\examples{
# sample of calculating subjects' accumulated DDDs
accumulat_DDD <- calDDDsAccumulation(case = sample_data_subset_atc, PatientIdColName = MemberId,DispensingColName = DispenseDate,AtcCodeColName = ATC,DailyDosageColName = DailyDosage,DurationColName = DaysSupply)
head(accumulat_DDD)
}
