% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDDDsRange.R
\name{calDDDsRange}
\alias{calDDDsRange}
\title{Calculate subjects' accumulated DDDs in a custom interval period}
\usage{
calDDDsRange(
  case,
  index_dayColName = Index_Day,
  expo_range_before = 36000,
  expo_range_after = 36000,
  idColName = MemberId,
  AtcCodeColName = ATC,
  DispenseDateColName = DispenseDate,
  DaysSupplyColName = DaysSupply,
  DailyDosageColName = DailyDosage
)
}
\arguments{
\item{case}{data.frame include subjects' id, dispensing date, drug ATC code, daily dosage, duration}

\item{expo_range_before}{days before observation day}

\item{expo_range_after}{days after observation day}

\item{idColName}{a colum for subject's id}

\item{DispenseDateColName}{a colum for dispensing}

\item{DaysSupplyColName}{a colum for supply day}

\item{DailyDosageColName}{a colum for daily dosage}

\item{index_day}{observation day}
}
\description{
Calculate subjects' accumulated DDDs in a custom interval period
}
\details{
This function provides user to calculate the total amount of DDDs accumulated by the patient in a custom interval period or specific date.
}
\examples{
# sample of calculating subjects' accumulated DDDs in a custom interval period
index_DDD <<- calDDDsRange(case = sample_data_subset_atc,index_dayColName = Index_Day,expo_range_before = 15,expo_range_after = 30,idColName = MemberId,AtcCodeColName = ATC,DispenseDateColName = DispenseDate,DaysSupplyColName = DaysSupply,DailyDosageColName = DailyDosage)
head(index_DDD)
}
