% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDDDs.R
\docType{data}
\name{ddd.test.df}
\alias{ddd.test.df}
\title{Calculate subjects' ddd}
\format{An object of class \code{data.frame} with 5 rows and 5 columns.}
\usage{
ddd.test.df
}
\arguments{
\item{case}{data.frame include subjects' id, dispensing date, drug ATC code, daily dosage, duration}
}
\description{
Calculate subjects' ddd
}
\keyword{datasets}
