% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATC1LevelPlot.R
\name{get.ATC1LevelPlot}
\alias{get.ATC1LevelPlot}
\title{Get ATC 1 level histogram plot}
\usage{
get.ATC1LevelPlot(df, ATCColName = ATC)
}
\arguments{
\item{df}{data.frame include ATC code}

\item{ATCColName}{A colum for ATC of df}
}
\description{
Get ATC 1 level histogram plot
}
