% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getATCBasedOnNHINo.R
\name{get.AtcViaNHINo}
\alias{get.AtcViaNHINo}
\title{Get ATC Code based on NHINO}
\usage{
get.AtcViaNHINo(df, NHINoColName = NHINo)
}
\arguments{
\item{df}{data.frame include NHINO}

\item{NHINoColName}{A colum for NHINo of df}

\item{Differ_ColName}{if colum for NHINo of df not named "NHINo"}
}
\description{
Get ATC Code based on NHINO
}
