% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getATCBaseOnRxCui.R
\name{get.AtcViaRxCui}
\alias{get.AtcViaRxCui}
\title{Get NDC code based on RxCui
WHO Collaborating Centre for Drug Statistics Methodology, ATC classification index with DDDs, 2019. Oslo, Norway 2018}
\usage{
get.AtcViaRxCui(df, RxCuiColName = RxCui, cores = 4,
  MatchRoute = TRUE)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{RxCuiColName}{A colum for RxCui of df}

\item{cores}{number of parallel operation}

\item{MatchRoute}{An option to mapping ingredient and route}
}
\description{
Get NDC code based on RxCui
WHO Collaborating Centre for Drug Statistics Methodology, ATC classification index with DDDs, 2019. Oslo, Norway 2018
}
