% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispensingPlot.R
\name{get.DispensingPlot}
\alias{get.DispensingPlot}
\title{Get dispensing plot}
\usage{
get.DispensingPlot(df, MemberIDColName = MemberID, Member,
  DrugColName = NationalDrugCode, DispenseDateColName = Dispensing,
  DaysSupplyColName = DaysSupply, TimeInterval = 1, Unit = month)
}
\arguments{
\item{df}{data.frame include dispense info}

\item{MemberIDColName}{A colum for memebr id}

\item{Member}{A colum for the member who be plot}

\item{DrugColName}{A colum for drug name}

\item{DispenseDateColName}{A colum for dispense date}

\item{DaysSupplyColName}{A colum for day supplys}

\item{TimeInterval}{A colum for time interval}

\item{Unit}{A colum for time unit}
}
\description{
Get dispensing plot
}
