% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.DrugEra.R
\name{get.DrugEra}
\alias{get.DrugEra}
\title{Generate Drug Era}
\usage{
get.DrugEra(df, window = 30, MemberIDColName = MemberID,
  DrugColName = Drug, DispenseDateColName, DaysSupplyColName,
  StartDateColName, EndDateColName)
}
\arguments{
\item{df}{data.frame with MemberID ,Drug,DispenseDate ,DaysSupply, or with MemberID ,Drug, StartDate, EndDate}

\item{window}{allowed gap between pharmacy claims, default is 30}

\item{DrugColName}{A colum for drug which patient use}

\item{DispenseDateColName}{A colum for dispense date}

\item{DaysSupplyColName}{A colum for drug supply days}

\item{StartDateColName}{A colum for drug start day}

\item{EndDateColName}{A colum for drug end day}
}
\description{
This can be used to merge pharmacy claims data into drug era with defined window. Exposure days will be cacluated, too.
}
