% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugEraPlot.R
\name{get.DrugEraPlot}
\alias{get.DrugEraPlot}
\title{Get drug histogram plot}
\usage{
get.DrugEraPlot(df, MemberIDColName = MemberID, DrugColName,
  DrugEraColName = DrugEra, SupplyDaysColName = SupplyDays)
}
\arguments{
\item{df}{data.frame include drgu era info}

\item{MemberIDColName}{A colum for member id of df}

\item{DrugColName}{A colum for drug og df}

\item{DrugEraColName}{A colum for drug era of df}

\item{SupplyDaysColName}{A colum for supply days fo df}
}
\description{
Get drug histogram plot
}
