% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMayTreat.R
\name{get.MayTreat}
\alias{get.MayTreat}
\title{Get may treat based on RxCui}
\usage{
get.MayTreat(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{RxCuiColName}{A colum for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
Get may treat based on RxCui
}
