% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMEDRTinfo.R
\name{get.MEDRTinfo}
\alias{get.MEDRTinfo}
\title{Get the Veterans Health Administration’s Medication Reference Terminology (MED-RT) information based on RxCui}
\usage{
get.MEDRTinfo(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{RxCuiColName}{A colum for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
Get the Veterans Health Administration’s Medication Reference Terminology (MED-RT) information based on RxCui
}
