% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNDCBasedOnSBDrxcui.R
\name{get.NdcViaSBDrxcui}
\alias{get.NdcViaSBDrxcui}
\title{Get NDC code based on RxCui}
\usage{
get.NdcViaSBDrxcui(df, SBDRxCuiColName = SBD.rxcui, cores = 8)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{cores}{number of parallel operation}

\item{RxCuiColName}{A colum for RxCui of df}
}
\description{
Get NDC code based on RxCui
}
