% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.NHINoBasedOnHC.R
\name{get.NHINoViaHC}
\alias{get.NHINoViaHC}
\title{Get NHINO based on RCFNO}
\usage{
get.NHINoViaHC(df, HospitalCodeColName, SourceDf, Source_NhinoColName,
  Source_HospitalCodeColName)
}
\arguments{
\item{df}{data.frame include specific drug code}

\item{HospitalCodeColName}{A colum for specific drug code}

\item{SourceDf}{data.frame include mapping between specific drug code and NHINo}

\item{Source_NhinoColName}{A colum for NHIno in SourceDf}

\item{Source_HospitalCodeColName}{A colum for specific drug code in SourceDf}
}
\description{
Get NHINO based on RCFNO
}
