% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNHINoBasedOnRCFNo.R
\name{get.NHINoViaRCFNo}
\alias{get.NHINoViaRCFNo}
\title{Get NHINO based on RCFNO}
\usage{
get.NHINoViaRCFNo(df, RCFNoColName = RCFNO)
}
\arguments{
\item{df}{data.frame include RCFNO}

\item{RCFNoColName}{A colum for RCFNo of df}
}
\description{
Get NHINO based on RCFNO
}
