% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxCuiBasedOnATC.R
\name{get.RxCuiViaAtc}
\alias{get.RxCuiViaAtc}
\title{Get RxCui based on ATC}
\usage{
get.RxCuiViaAtc(df, AtcColName = ATC, cores = 8)
}
\arguments{
\item{df}{data.frame include ATC}

\item{AtcColName}{A colum for ATC of df}

\item{cores}{number of parallel operation}

\item{Differ_ColName}{if colum for NHINo of df not named "ATC"}
}
\description{
Get RxCui based on ATC
}
