% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxCuiBasedOnMayprevent.R
\name{get.RxCuiViaMayprevent}
\alias{get.RxCuiViaMayprevent}
\title{Get RxCui based on may prevent}
\usage{
get.RxCuiViaMayprevent(strmayprevent)
}
\arguments{
\item{strmaytreat}{A may prevent}
}
\description{
Get RxCui based on may prevent
}
