% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxCuiBasedOnMaytreat.R
\name{get.RxCuiViaMaytreat}
\alias{get.RxCuiViaMaytreat}
\title{Get RxCui based on may treat}
\usage{
get.RxCuiViaMaytreat(strmaytreat)
}
\arguments{
\item{strmaytreat}{A may treat}
}
\description{
Get RxCui based on may treat
}
