% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxcuiBasedOnNDC.R
\name{get.RxCuiViaNdc}
\alias{get.RxCuiViaNdc}
\title{Get RxCui based on NDC}
\usage{
get.RxCuiViaNdc(df, NdcColName = NDC, cores = 8)
}
\arguments{
\item{df}{data.frame include NDC}

\item{AtcColName}{A colum for NDC of df}

\item{Differ_ColName}{if colum for NHINo of df not named "NDC"}
}
\description{
Get RxCui based on NDC
}
