% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxInfo.R
\name{get.rxinfo}
\alias{get.rxinfo}
\title{Get RxCui infomation}
\usage{
get.rxinfo(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{RxCuiColName}{A colum for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
Get RxCui infomation
}
