% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtcLevelPlot.R
\name{getAtcLevelPlot}
\alias{getAtcLevelPlot}
\title{Get ATC level histogram plot}
\usage{
getAtcLevelPlot(df, ATCColName = ATC, level = 1)
}
\arguments{
\item{df}{data.frame include ATC code}

\item{ATCColName}{A colum for ATC of df}
}
\description{
Get ATC level histogram plot
}
\details{
This function provides user to get Anatomical Therapeutic Chemical Classification System(ATC) level 1 or 2 histogram plot.
}
\examples{
# sample of getting ATC level 1 histogram plot.
getAtcLevelPlot(df = sample_data_ATC1LevelPlot, ATCColName = ATC, level = 1)
# sample of getting ATC level 2 histogram plot.
getAtcLevelPlot(df = sample_data_ATC1LevelPlot, ATCColName = ATC, level = 2)
}
