% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getATCLevelPlotTest1.R
\name{getATCLevelPlotTest1}
\alias{getATCLevelPlotTest1}
\alias{getATCLevelPlot}
\title{Get  Anatomical Therapeutic Chemical Classification System(ATC) code level code distribution histogram plot}
\usage{
getATCLevelPlot(df, ATCColName = ATC, level = 1)
}
\arguments{
\item{df}{data.frame, include ATC code}

\item{ATCColName}{a column name for ATC code of df}
}
\description{
\code{getAtcLevelPlot} is a function to get daily dosage drugs(DDDs) in Anatomical Therapeutic Chemical Classification System(ATC) code level histogram plot.
}
\details{
If the amount of drug information obtained by the user is large and messy, the user can first visualize the data using basic statistical tools, explore the characteristics of the data, and obtain the information, structure and features contained in the data.
Because before having a complicated or rigorous analysis, user must have more knowledge of the data to determine the direction of the data analysis.
}
\examples{
# sample of getting ATC first-level code distribution bar chart
getATCLevelPlot(df = sample_data_ATC1LevelPlot, ATCColName = ATC, level = 1)
# sample of getting ATC second-level code distribution bar chart
getATCLevelPlot(df = sample_data_ATC1LevelPlot, ATCColName = ATC, level = 2)
}
