% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getAtcViaNHINo.R
\name{getAtcViaNHINo}
\alias{getAtcViaNHINo}
\title{Get ATC via NHINO}
\usage{
getAtcViaNHINo(df, NHINoColName = NHINo)
}
\arguments{
\item{df}{data.frame include NHINO}

\item{NHINoColName}{A colum for NHINo of df}
}
\description{
\code{getAtcViaNHINo} is a function to get Anatomical Therapeutic Chemical Classification System(ATC) via Taiwan Health Insurance drug code(NHINo)
}
\details{
This function provides user to get Anatomical Therapeutic Chemical Classification System(ATC) via Taiwan Health Insurance drug code(NHINo).
}
\examples{
# sample of getting ATC via NHINo.
head(getAtcViaNHINo(df = sample_nhino_code,NHINoColName = NHINo))
}
