% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getAtcViaRxcui.R
\name{getAtcViaRxcui}
\alias{getAtcViaRxcui}
\title{Get ATC via RxCui}
\usage{
getAtcViaRxcui(df, RxCuiColName = RxCui, cores = 4, MatchRoute = FALSE)
}
\arguments{
\item{df}{data.frame, a drug ingredient`s RxCui of the specific indication}

\item{cores}{number of parallel operation}

\item{MatchRoute}{an option to mapping drug ingredient and route}

\item{RxCuiColNamea}{a column name for RxCui of df}
}
\value{
A new data.table \code{sample_data_subset_atc} with following columns.

\code{MemberId} patient Id

\code{DispenseDate} patient`s drug dispensed date, and the time interval is from 2025-12-06 to 2038-01-25

\code{NationalDrugCode} drug NDC, total of 17,326 different drug codes

\code{Quantity} quantity of patient`s drug dispensed

\code{DaysSupply} days of patient`s drug dispensed

\code{RxCui} drug RxCui

\code{ATC} drug ATC
}
\description{
\code{getRxcuiViaNdc} is a function to get Anatomical Therapeutic Chemical Classification System(ATC) via RxCui
}
\details{
WHO Collaborating Centre for Drug Statistics Methodology, ATC classification index with DDDs, 2019. Oslo, Norway 2018.
The same drug components in ATC might depend on different ATC medication route.
Hence, this function also  refer to the research of \code{Bodenreider et al.}(https://www.researchgate.net/publication/276067397_Analyzing_US_prescription_lists_with_RxNorm_and_the_ATCDDD_Index) and adds the parameter MatchRoute for user to consider.
If user wants to consider and compare the ATC medication route, input \code{TRUE} in the parameter \code{MatchRoute}, otherwise input \code{FALSE}.
This function provides user to get ATC via RxCui.
}
\examples{
# sample of getting ATC via RxCui.
sample_data_subset_atc <- getAtcViaRxcui(df = sample_data_subset_rxcui,RxCuiColName = RxCui,cores = 2,MatchRoute = FALSE)
head(sample_data_subset_atc)
}
