% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoSSViaRxcui.R
\name{getBoSSViaRxcui}
\alias{getBoSSViaRxcui}
\title{Get Ingredient and Basis of strength substance (BoSS)}
\usage{
getBoSSViaRxcui(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{RxCuiColName}{A colum for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
Get Ingredient and Basis of strength substance (BoSS)
}
