% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getDispensingPlot.R
\name{getDispensingPlot}
\alias{getDispensingPlot}
\title{Get specific medication history dispensing plot}
\usage{
getDispensingPlot(
  df,
  MemberIDColName = MemberID,
  Member,
  DrugColName = NationalDrugCode,
  DispenseDateColName = Dispensing,
  DaysSupplyColName = DaysSupply,
  TimeInterval = 1,
  Unit = month
)
}
\arguments{
\item{df}{data.frame, include dispensed info}

\item{MemberIDColName}{a column name for member id of df}

\item{Member}{a column name for the member name of plot}

\item{DrugColName}{a column name for drug name of df}

\item{DispenseDateColName}{a column name for dispensed date of df}

\item{DaysSupplyColName}{a column name for day supplied of df}

\item{TimeInterval}{a column name for time interval of df}

\item{Unit}{a column name for time unit of df}
}
\description{
This function provides user to select specific patient to visualize their medication history in a custom interval period or specific date.
}
\details{
In addition to visualize the results of the drug era for all patients in the data, user can also select a single patient to see the medication history of the patient's specific drug, and mark the length of each medication record and the time interval between each medication record.
}
\examples{
#sample of getting dispensing plot.
getDispensingPlot(df = sample_data_subset, MemberIDColName = MemberId,DrugColName = NationalDrugCode,DispenseDateColName = Dispensing,Member = 42, TimeInterval = 20, Unit = day)
}
