% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getDrugEraPlot.R
\name{getDrugEraPlot}
\alias{getDrugEraPlot}
\title{Get drug era histogram plot}
\usage{
getDrugEraPlot(
  df,
  MemberIDColName = MemberID,
  DrugColName,
  DrugEraColName = DrugEra,
  SupplyDaysColName = SupplyDays
)
}
\arguments{
\item{df}{data.frame include drug era info}

\item{MemberIDColName}{a column name for member id of df}

\item{DrugColName}{a column name for drug name df}

\item{DrugEraColName}{a column name for drug era of df}

\item{SupplyDaysColName}{A column name for supply days of df}
}
\description{
This function provides user to get drug era histogram plot.
}
\details{
User can get the drug data, the length of each drug record, and the number of days of drug supply in the drug era.
In addition to the analysis of each drug route by the drug era, user can also visualize the drug data of the drug era.
}
\examples{
#sample of getting drug era histogram plot.
getDrugEraPlot(df = sample_data_DrugEraPlot,MemberIDColName = MemberId,DrugColName = ATC,DrugEraColName = DrugEra,SupplyDaysColName = SupplyDays)
}
