% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getMayPrevent.R
\name{getMayPrevent}
\alias{getMayPrevent}
\title{Get specific indication may prevent via RxCui}
\usage{
getMayPrevent(df, RxCuiColName = RxCui, cores = 16)
}
\arguments{
\item{df}{data.frame, a drug ingredient RxCui}

\item{RxCuiColName}{a column name for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
This function provides user to get the specific indication may prevent via drug RxCui.
}
