% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getMEDRTInfo.R
\name{getMEDRTInfo}
\alias{getMEDRTInfo}
\title{Get the Veterans Health Administration's Medication Reference Terminology (MED-RT) information via RxCui}
\usage{
getMEDRTInfo(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame, included RxCui}

\item{RxCuiColName}{a column name for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
This function can get \code{MED-RT} information via RxCui.
}
