% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getNdcViaSbdRxcui.R
\name{getNdcViaSbdRxcui}
\alias{getNdcViaSbdRxcui}
\title{Get NDC via SBD RxCui}
\usage{
getNdcViaSbdRxcui(df, SBDRxCuiColName = SBD.rxcui, cores = 8)
}
\arguments{
\item{df}{data.frame, a drug ingredient`s SBD or SCD RxCui of the specific indication}

\item{cores}{number of parallel operation}

\item{RxCuiColName}{a column name for RxCui of df}
}
\value{
A new data.table \code{maytreat_NDC_List} with following columns.

\code{RxCui} drug RxCui

\code{Name} drug name

\code{MinConcept.Id} indication Id

\code{MinConcept.Name} indication name

\code{SBD.rxcui} drug SBD RxCui

\code{NDC} drug NDC
}
\description{
\code{getNdcViaSbdRxcui} is a function to get National Drug Code(NDC) via Semantic Branded Drug(SBD) RxCui or Semantic Clinical Drug(SCD) RxCui
}
\details{
After obtaining drug ingredient`s SBD RxCui of the specific indication, user can get drug NDC contained these drug ingredients. However, a drug will have different packaging and NDC because it comes from different pharmaceutical brands. There will be different NDC for the same drug type but different packaging. Hence, the output results will be one-to-many. A drug RxCui will matche to multiple NDC of pharmaceutical brands, and no corresponding NDC will be \code{NA}. This function provides user to get NDC via SBD or SCD RxCui.
}
\examples{
# sample of getting NDC via SBD SCD RxCui list of esophagitis
Esophagitis_NDC_List <- getNdcViaSbdRxcui(df = Esophagitis_SBD_SCD_List[1,],SBDRxCuiColName = SBD.rxcui,cores = 2)
head(Esophagitis_NDC_List)
}
