% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getNHINoViaHC.R
\name{getNHINoViaHC}
\alias{getNHINoViaHC}
\title{Get NHINo via HC}
\usage{
getNHINoViaHC(
  df,
  HospitalCodeColName,
  SourceDf,
  Source_NhinoColName,
  Source_HospitalCodeColName
)
}
\arguments{
\item{df}{data.frame, the specific hospital drug code}

\item{HospitalCodeColName}{a column name for a specific hospital drug code of df}

\item{SourceDf}{data.frame include mapping between specific hospital drug code and NHINo}

\item{Source_NhinoColName}{a column name for NHINo in Sourcedf}

\item{Source_HospitalCodeColName}{a column name for specific hospital drug code in Sourcedf}
}
\value{

}
\description{
\code{getNHINoViaHC} is a function to get Taiwan Health Insurance drug code(NHINo) via Hospital Code(HC)
}
\details{
This function provides user to get NHINo via HC.
}
