% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNHINoViaRCFNo.R
\name{getNHINoViaRCFNo}
\alias{getNHINoViaRCFNo}
\title{Get NHINO based on RCFNO}
\usage{
getNHINoViaRCFNo(df, RCFNoColName = RCFNO)
}
\arguments{
\item{df}{data.frame include RCFNO}

\item{RCFNoColName}{A colum for RCFNo of df}
}
\description{
Get NHINO based on RCFNO
}
\details{
This function provides user to get Taiwan Health Insurance drug code(NHINo) via Taiwan hospital drug code(RCFNo).
}
\examples{
# sample of getting NHINo via RCFNo of Chang Gung Medical Hospital.
head(getNHINoViaRCFNo(df = sample_taiwan_drug, RCFNoColName = CGMH_CODE))
}
