% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxcuiViaAtc.R
\name{getRxcuiViaAtc}
\alias{getRxcuiViaAtc}
\title{Get RxCui based on ATC}
\usage{
getRxcuiViaAtc(df, AtcColName = ATC, cores = 8)
}
\arguments{
\item{df}{data.frame include ATC}

\item{AtcColName}{A colum for ATC of df}

\item{cores}{number of parallel operation}

\item{Differ_ColName}{if colum for NHINo of df not named "ATC"}
}
\description{
Get RxCui based on ATC
}
\details{
This function provides user to get RxCui via Anatomical Therapeutic Chemical Classification System(ATC).
}
\examples{
# sample of getting RxCui via ATC.
sample_atc_rxcui <- getRxcuiViaAtc(df = sample_ATC,AtcColName = ATC,cores = 2)
head(sample_data_subset_atc)
}
