% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getRxCuiViaMayPreventTest1.R
\name{getRxCuiViaMayPrevent}
\alias{getRxCuiViaMayPrevent}
\title{Get ingredient RxCui via specific indication}
\usage{
getRxCuiViaMayPrevent(strmayprevent)
}
\arguments{
\item{strmayprevent}{user can input an indication name}
}
\description{
This function provides user to get drug ingredient RxCui of the specific indication may prevent.
}
