% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getRxCuiViaMayPrevent123.R
\name{getRxCuiViaMayPrevent123}
\alias{getRxCuiViaMayPrevent123}
\title{Get ingredient RxCui via specific indication}
\usage{
getRxCuiViaMayPrevent123(strmayprevent)
}
\arguments{
\item{strmayprevent}{user can input an indication name}
}
\description{
This function provides user to get drug ingredient RxCui of the specific indication may prevent.
}
