% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxcuiViaMaytreat.R
\name{getRxcuiViaMaytreat}
\alias{getRxcuiViaMaytreat}
\title{Get RxCui based on may treat}
\usage{
getRxcuiViaMaytreat(strmaytreat)
}
\arguments{
\item{strmaytreat}{A may treat}
}
\description{
Get RxCui based on may treat
}
\details{
This function provides user to get medicine ingredient RxCui via specific indication Maytreat
}
\examples{
# sample of searching an indication - esophagitis.
Esophagitis_List <- getRxcuiViaMaytreat("esophagitis")
# sample of getting RxCui list of esophagitis.
head(Esophagitis_List)
}
