% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getRxcuiViaNdc.R
\name{getRxcuiViaNdc}
\alias{getRxcuiViaNdc}
\title{Get RxCui via NDC}
\usage{
getRxcuiViaNdc(df, NdcColName = NDC, cores = 8)
}
\arguments{
\item{df}{data.frame, a drug ingredient`s NDC of the specific indication}

\item{NdcColName}{a column name for NDC of df}

\item{cores}{number of parallel operation}
}
\value{
A new data.table \code{sample_data_subset_rxcui} with following columns.

\code{MemberId} patient Id

\code{DispenseDate} patient`s drug dispensed date, and the time interval is from 2025-12-06 to 2038-01-25

\code{NationalDrugCode} drug NDC, total of 17,326 different drug codes

\code{Quantity} quantity of patient`s drug dispensed

\code{DaysSupply} days of patient`s drug dispensed

\code{RxCui} drug RxCui

\code{ndcStatus} drug NDC status??
}
\description{
\code{getRxcuiViaNdc} is a function to get RxCui via National Drug Code(NDC)
}
\details{
After obtaining \code{sample_data_subset}, user can get drug RxCui contained these drug ingredients. This function provides user to get RxCui via NDC.
}
\examples{
#sample of getting RxCui via NDC.
sample_data_subset_rxcui <- getRxcuiViaNdc(df = sample_data_subset,NdcColName = NationalDrugCode)
head(sample_data_subset_rxcui)
}
