% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getSbdRxcuiViaRxcui.R
\name{getSbdRxcuiViaRxcui}
\alias{getSbdRxcuiViaRxcui}
\title{Get SBD RxCui via RxCui}
\usage{
getSbdRxcuiViaRxcui(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame, a drug ingredient`s RxCui of the specific indication}

\item{RxCuiColName}{a column name for RxCui of df}

\item{cores}{number of parallel operation}
}
\value{
A new data.table \code{maytreat_SBD_SCD_List} with following columns.

\code{RxCui} drug RxCui

\code{Name} drug name

\code{MinConcept.Id} indication Id

\code{MinConcept.Name} indication name

\code{SBD.rxcui} drug SBD RxCui
}
\description{
\code{getSbdRxcuiViaRxcui} is a function to get Semantic Branded Drug(SBD) or Semantic Clinical Drug(SCD) RxCui via RxCui
}
\details{
After obtaining drug ingredient`s RxCui of the specific indication, user can get all pharmaceutical brand and clinical drug contained these drug ingredients. This function provides user to get SBD or SCD RxCui via ingredient RxCui.
}
\examples{
# sample of getting SBD or SCD Rxcui via RxCui list of esophagitis.
Esophagitis_SBD_SCD_List <- getSbdRxcuiViaRxcui(df = Esophagitis_List[3,],RxCuiColName = RxCui,cores = 2)
head(Esophagitis_SBD_SCD_List)
}
