% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getSBDRxCuiViaRxCui123.R
\name{getSBDRxCuiViaRxCui123}
\alias{getSBDRxCuiViaRxCui123}
\title{Get Semantic Branded Drug(SBD) or Semantic Clinical Drug(SCD) RxCui via RxCui}
\usage{
getSBDRxCuiViaRxCui123(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame, a drug ingredient RxCui of the specific indication}

\item{RxCuiColName}{a column name for RxCui of df}

\item{cores}{number of parallel operation}
}
\value{
A new data.table \code{maytreat_SBD_SCD_List} with following columns.

\code{RxCui} drug RxCui

\code{Name} drug name

\code{MinConcept.Id} indication Id

\code{MinConcept.Name} indication name

\code{SBD.rxcui} drug SBD RxCui
}
\description{
This function provides user to get Semantic Branded Drug(SBD) or Semantic Clinical Drug(SCD) RxCui via ingredient RxCui.
}
\details{
After obtaining the data.frame from \code{getRxCuiViaMayTreat()} and \code{getRxCuiViaMayPrevent()}, user can further obtain a commercially available pharmaceutical brand and clinical drug contained these drug ingredients.
The content is to add a column named \code{SBD.rxcui} in the original input data frame, stores the result of converting ingredient to brand drug and clinical drug containing a specific drug component.
}
\examples{
# sample of getting SBD or SCD Rxcui list of esophagitis.
Esophagitis_SBD_SCD_List <- getSBDRxCuiViaRxCui123(df = Esophagitis_List[3,],RxCuiColName = RxCui,cores = 2)
head(Esophagitis_SBD_SCD_List)
}
