% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDDDs_accumulat.R
\name{calDDDs.accumulat}
\alias{calDDDs.accumulat}
\title{Calculate subjects' acculated ddd before last dispensing}
\usage{
calDDDs.accumulat(case)
}
\arguments{
\item{case}{data.frame include subjects' id, dispensing date, drug ATC code, daily dosage, duration}
}
\description{
Calculate subjects' acculated ddd before last dispensing
}
