% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDDDsAccumulation.R, R/funPharm.R,
%   R/getDDDs.R
\name{calDDDsAccumulation}
\alias{calDDDsAccumulation}
\alias{getDDDs}
\title{Calculate subjects' total daily dosage drugs(DDDs)}
\usage{
calDDDsAccumulation(
  case,
  PatientIdColName = MemberId,
  DispensingColName = DispenseDate,
  AtcCodeColName = ATC,
  DailyDosageColName = DailyDosage,
  DurationColName = DaysSupply
)

getDDDs(atc)
}
\arguments{
\item{case}{data.frame, include subjects' id, dispensing date, drug ATC code, daily dosage, duration}

\item{atc}{data.frame include ATC code}
}
\value{
A new data.table \code{accumulat_DDD} with following columns.

\code{MemberId} patient Id

\code{DDDs} accumulated DDDs
}
\description{
\code{calDDDsAccumulation} is a function to calculate subjects' accumulated daily dosage drugs(DDDs) in RxCui before last dispensed.
\code{getDDDs} is a function to get DDDs in ATC code
}
\details{
User can input data frame containing the patient id, dispensing date, the drug ATC code, drug supply days, and daily dose of the drug column.
This function will output a data frame which store DDDs accumulated by patients.
}
\examples{
# sample of calculating subjects' accumulated DDDs
accumulat_DDD <- calDDDsAccumulation(case = sample_data_subset_atc, PatientIdColName = MemberId,DispensingColName = DispenseDate,AtcCodeColName = ATC,DailyDosageColName = DailyDosage,DurationColName = DaysSupply)
head(accumulat_DDD)
}
