% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDDDsAccumulation.R
\name{calDDDsAccumulation}
\alias{calDDDsAccumulation}
\title{Calculate subjects' acculated ddd before last dispensing}
\usage{
calDDDsAccumulation(
  case,
  PatientIdColName = MemberId,
  DispensingColName = DispenseDate,
  AtcCodeColName = ATC,
  DailyDosageColName = DailyDosage,
  DurationColName = DaysSupply
)
}
\arguments{
\item{case}{data.frame include subjects' id, dispensing date, drug ATC code, daily dosage, duration}
}
\description{
Calculate subjects' acculated ddd before last dispensing
}
\details{
This function provides user to calculate patient`s total amount of DDDs for accumulation medication time.
}
\examples{
# sample of calculating subjects' accumulated DDDs
accumulat_DDD <<- calDDDsAccumulation(case = sample_data_subset_atc, PatientIdColName = MemberId,DispensingColName = DispenseDate,AtcCodeColName = ATC,DailyDosageColName = DailyDosage,DurationColName = DaysSupply)
head(accumulat_DDD)
}
