% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDDDsRange.R, R/funPharm.R
\name{calDDDsRange}
\alias{calDDDsRange}
\title{Calculate subjects' accumulated daily dosage drugs(DDDs) in custom interval period}
\usage{
calDDDsRange(
  case,
  index_dayColName = Index_Day,
  expo_range_before = 36000,
  expo_range_after = 36000,
  idColName = MemberId,
  AtcCodeColName = ATC,
  DispenseDateColName = DispenseDate,
  DaysSupplyColName = DaysSupply,
  DailyDosageColName = DailyDosage
)
}
\arguments{
\item{case}{data.frame include subjects' id, dispensing date, drug ATC code, daily dosage, duration}

\item{expo_range_before}{days before observation day}

\item{expo_range_after}{days after observation day}

\item{idColName}{a column name for subject's id}

\item{DispenseDateColName}{a column name for dispensing}

\item{DaysSupplyColName}{a column name for supply day}

\item{DailyDosageColName}{a column name for DDDs}

\item{index_day}{observation day}
}
\value{
A new data.table \code{index_DDD} with following columns.

\code{MemberId} patient Id

\code{Start_day} the begin date of patient take medicine

\code{Index_Day} the begin date of calculating the DDDs

\code{End_day} the end date of patient take medicine

\code{DDDs_Before_15_Days} calculate DDDs in custom time interval before index 15 day

\code{DDDs_After_30_Days} calculate DDDs in custom time interval after index 30 day
}
\description{
\code{calDDDsRange} is a function to calculate subjects' accumulated daily dosage drugs(DDDs) in RxCui in custom interval period
\code{getDDDs} is a function to get DDDs in ATC code
}
\details{
User can get the total amount of DDDs accumulated of patients before and after a specific date in a custom interval period or specific date.
}
\examples{
# sample of calculating subjects' accumulated DDDs in a custom interval period
index_DDD <- calDDDsRange(case = sample_data_subset_atc,index_dayColName = Index_Day,expo_range_before = 15,expo_range_after = 30,idColName = MemberId,AtcCodeColName = ATC,DispenseDateColName = DispenseDate,DaysSupplyColName = DaysSupply,DailyDosageColName = DailyDosage)
head(index_DDD)
}
