% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxcuiBasedOnNDC.R
\name{get.RxCuiViaNdc}
\alias{get.RxCuiViaNdc}
\title{Get RxCui based on NDC}
\usage{
get.RxCuiViaNdc(df, NdcColName = NDC, cores = 8)
}
\arguments{
\item{df}{data.frame include NDC}

\item{NdcColName}{A colum for NDC of df}

\item{cores}{number of parallel operation}
}
\description{
Get RxCui based on NDC
}
