% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getATCViaNHINoTest1.R
\name{getATCViaNHINo}
\alias{getATCViaNHINo}
\title{Get Anatomical Therapeutic Chemical Classification System(ATC) code via Taiwan Health Insurance drug code(NHINo)}
\usage{
getATCViaNHINo(df, NHINoColName = NHINo)
}
\arguments{
\item{df}{data.frame include NHINO}

\item{NHINoColName}{A colum for NHINo of df}
}
\value{

}
\description{
This is a function to get Anatomical Therapeutic Chemical Classification System(ATC) code via Taiwan Health Insurance drug code(NHINo)
}
\details{
This function filter the patient to be analyzed with the target indication drug, which encode in RxCui and be converted to ATC code by the function \code{getATCViaRxCui()}.
User inputs the data frame with drug column which encoded in NHINo, then the output will add a column named ATC in the original input data frame.
}
\examples{
# sample of getting ATC code via NHINo.
head(getATCViaNHINo(df = sample_nhino_code,NHINoColName = NHINo))
}
