% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtcViaRxcui.R
\name{getAtcViaRxcui}
\alias{getAtcViaRxcui}
\title{Get NDC code based on RxCui
WHO Collaborating Centre for Drug Statistics Methodology, ATC classification index with DDDs, 2019. Oslo, Norway 2018}
\usage{
getAtcViaRxcui(df, RxCuiColName = RxCui, cores = 4, MatchRoute = FALSE)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{RxCuiColName}{A colum for RxCui of df}

\item{cores}{number of parallel operation}

\item{MatchRoute}{An option to mapping ingredient and route}
}
\description{
Get NDC code based on RxCui
WHO Collaborating Centre for Drug Statistics Methodology, ATC classification index with DDDs, 2019. Oslo, Norway 2018
}
\details{
This function provides user to get Anatomical Therapeutic Chemical Classification System(ATC) via RxCui. The same drug components in ATC might depend on different ATC medication route. Hence, this function also refers to the research of Bodenreider et al. and adds the parameter MatchRoute for user to consider. If user wants to consider and compare the ATC medication route, input TRUE in the parameter MatchRoute, otherwise input FALSE. # WHO Collaborating Centre for Drug Statistics Methodology, ATC classification index with DDDs, 2019. Oslo, Norway 2018
}
\examples{
# sample of getting ATC via RxCui.
sample_data_subset_atc <- getAtcViaRxcui(df = sample_data_subset_rxcui,RxCuiColName = RxCui,cores = 2,MatchRoute = FALSE)
head(sample_data_subset_atc)
}
