% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getATCViaRxCui.R
\name{getATCViaRxCui}
\alias{getATCViaRxCui}
\title{Get Anatomical Therapeutic Chemical Classification System(ATC) code via RxCui}
\usage{
getATCViaRxCui(df, RxCuiColName = RxCui, cores = 4, MatchRoute = FALSE)
}
\arguments{
\item{df}{data.frame, a drug ingredient`s RxCui of the specific indication}

\item{cores}{number of parallel operation}

\item{MatchRoute}{an option to mapping drug ingredient and route}

\item{RxCuiColNamea}{a column name for RxCui of df}
}
\value{
A new data.table \code{sample_data_subset_atc} with following columns.

\code{MemberId} patient Id

\code{DispenseDate} patient`s drug dispensed date, and the time interval is from 2025-12-06 to 2038-01-25

\code{NationalDrugCode} drug NDC, total of 17,326 different drug codes

\code{Quantity} quantity of patient`s drug dispensed

\code{DaysSupply} days of patient`s drug dispensed

\code{RxCui} drug RxCui

\code{ATC} drug ATC code
}
\description{
This is a function to get Anatomical Therapeutic Chemical Classification System(ATC) code via RxCui
}
\details{
User can input the data.frame with drug column which encoded in RxCui, and to add the result of converting column named ATC.

The same drug components in ATC might depend on different ATC medication route.
Hence, this function also refer to the research of \code{Bodenreider et al.}(https://www.researchgate.net/publication/276067397_Analyzing_US_prescription_lists_with_RxNorm_and_the_ATCDDD_Index) and adds the parameter \code{MatchRoute} for user to consider.
If user wants to compare the ATC medication route, input \code{TRUE} in the parameter \code{MatchRoute}, otherwise input \code{FALSE}.
Because RxCui is slightly different from the ATC drug vocabulary, not every RxCui has a corresponding ATC code.
Therefore, if RxCui does not have a corresponding ATC code, this field will display \code{NA}, and pharm does not provide the code conversion of the clinical drug and brand drug combination package, such as birth control pills.
If the input RxCui is a drug combination package, the field output result will be filled in GPCK or BPCK.
}
\examples{
# sample of getting ATC via RxCui.
sample_data_subset_atc <- getATCViaRxCui(df = sample_data_subset_rxcui,RxCuiColName = RxCui,cores = 2,MatchRoute = FALSE)
head(sample_data_subset_atc)
}
