% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getBoSSViaRxCui.R
\name{getBoSSViaRxCui}
\alias{getBoSSViaRxCui}
\title{Get Ingredient and Basis of Strength Substance (BoSS) information via RxCui}
\usage{
getBoSSViaRxCui(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame, included RxCui}

\item{RxCuiColName}{a column name for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
This function can get BoSS information.
}
