% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDispensingPlot.R
\name{getDispensingPlot}
\alias{getDispensingPlot}
\title{Get dispensing plot}
\usage{
getDispensingPlot(
  df,
  MemberIDColName = MemberID,
  Member,
  DrugColName = NationalDrugCode,
  DispenseDateColName = Dispensing,
  DaysSupplyColName = DaysSupply,
  TimeInterval = 1,
  Unit = month
)
}
\arguments{
\item{df}{data.frame include dispense info}

\item{MemberIDColName}{A colum for memebr id}

\item{Member}{A colum for the member who be plot}

\item{DrugColName}{A colum for drug name}

\item{DispenseDateColName}{A colum for dispense date}

\item{DaysSupplyColName}{A colum for day supplys}

\item{TimeInterval}{A colum for time interval}

\item{Unit}{A colum for time unit}
}
\description{
Get dispensing plot
}
\details{
This function provides user to selecting specific patients may visualize their medication history in a custom interval period or specific date.
}
\examples{
#sample of getting dispensing plot.
getDispensingPlot(df = sample_data_subset, MemberIDColName = MemberId,DrugColName = NationalDrugCode,DispenseDateColName = Dispensing,Member = 42, TimeInterval = 20, Unit = day)
}
