% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugEraPlot.R
\name{getDrugEraPlot}
\alias{getDrugEraPlot}
\title{Get drug histogram plot}
\usage{
getDrugEraPlot(
  df,
  MemberIDColName = MemberID,
  DrugColName,
  DrugEraColName = DrugEra,
  SupplyDaysColName = SupplyDays
)
}
\arguments{
\item{df}{data.frame include drgu era info}

\item{MemberIDColName}{A colum for member id of df}

\item{DrugColName}{A colum for drug og df}

\item{DrugEraColName}{A colum for drug era of df}

\item{SupplyDaysColName}{A colum for supply days fo df}
}
\description{
Get drug histogram plot
}
\details{
This function provides user to get drug era histogram plot.
}
\examples{
#sample of getting drug era histogram plot.
getDrugEraPlot(df = sample_data_DrugEraPlot,MemberIDColName = MemberId,DrugColName = ATC,DrugEraColName = DrugEra,SupplyDaysColName = SupplyDays)
}
