% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMayprevent.R
\name{getMayprevent}
\alias{getMayprevent}
\title{Get may prevent based on RxCui}
\usage{
getMayprevent(df, RxCuiColName = RxCui, cores = 16)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{RxCuiColName}{A colum for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
Get may prevent based on RxCui
}
