% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNdcViaSbdRxcui.R
\name{getNdcViaSbdRxcui}
\alias{getNdcViaSbdRxcui}
\title{Get NDC code based on RxCui}
\usage{
getNdcViaSbdRxcui(df, SBDRxCuiColName = SBD.rxcui, cores = 8)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{cores}{number of parallel operation}

\item{RxCuiColName}{A colum for RxCui of df}
}
\description{
Get NDC code based on RxCui
}
\details{
This function provides user to get National Drug Code(NDC) via Semantic Branded Drug(SBD) or Semantic Clinical Drug(SCD) RxCui. There will be different NDC for the same drug type but different packaging. Hence, the output NDC of this function will be one-to-many.
}
\examples{
# sample of getting NDC via SBD SCD RxCui list of esophagitis
Esophagitis_NDC_List <- getNdcViaSbdRxcui(df = Esophagitis_SBD_SCD_List[1,],SBDRxCuiColName = SBD.rxcui,cores = 2)
head(Esophagitis_NDC_List)
}
