% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getNDCViaSBDRxCui.R
\name{getNDCViaSBDRxCui}
\alias{getNDCViaSBDRxCui}
\title{Get U.S National Drug Code(NDC) via Semantic Branded Drug(SBD) or Semantic Clinical Drug(SCD) RxCui}
\usage{
getNDCViaSBDRxCui(df, SBDRxCuiColName = SBD.rxcui, cores = 8)
}
\arguments{
\item{df}{data.frame, a drug ingredient SBD or SCD RxCui of the specific indication}

\item{cores}{number of parallel operation}

\item{RxCuiColName}{a column name for RxCui of df}
}
\value{
A new data.table \code{maytreat_NDC_List} with following columns.

\code{RxCui} drug RxCui

\code{Name} drug name

\code{MinConcept.Id} indication Id

\code{MinConcept.Name} indication name

\code{SBD.rxcui} drug SBD RxCui

\code{NDC} drug NDC
}
\description{
This function provides user tO get U.S National Drug Code(NDC) via Semantic Branded Drug(SBD) RxCui or Semantic Clinical Drug(SCD) RxCui contained these drug ingredients.
}
\details{
User can screen out the patient who has the record of receiving these drugs in the medication data, but not all medication data are encoded in RxNorm.
In the United States, for example, the drug code commonly used in medical data is NDC code.
NDC code cannot be directly filtered with RxNorm coded data.
Therefore, user can input the data.frame from \code{getRxCuiViaMayTreat()}, then convert to NDC code and screen out a new data.frame.

However, a drug will have different packaging and NDC because it comes from different pharmaceutical brands.
There will be different NDC for the same drug type but different packaging. Hence, the output results will be one-to-many.
A drug RxCui will matche to multiple NDC of pharmaceutical brands, and no corresponding NDC will be \code{NA}.
}
\examples{
# sample of getting NDC via SBD SCD RxCui list of esophagitis
Esophagitis_NDC_List <- getNDCViaSBDRxCui(df = Esophagitis_SBD_SCD_List[1,],SBDRxCuiColName = SBD.rxcui,cores = 2)
head(Esophagitis_NDC_List)
}
