% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getNHINoViaRCFNo.R
\name{getNHINoViaRCFNo}
\alias{getNHINoViaRCFNo}
\title{Get Taiwan Health Insurance drug code(NHINo) via Hospital drug code(RCFNo)}
\usage{
getNHINoViaRCFNo(df, RCFNoColName = RCFNO)
}
\arguments{
\item{df}{data.frame, the specific hospital drug code}

\item{RCFNoColName}{a colum name for RCFNo of df}
}
\value{
A new data.table with following columns.

\code{CGMH_CODE} Chang Gung Medical Hospital drug code RCFNo

\code{NHINO1} drug NHINo
}
\description{
This is a function to get Taiwan Health Insurance drug code(NHINo) via Chang Gung Medical Hospital drug code(RCFNo)
}
\details{
User can input the specific hospital drug code
If there is no corresponed NHINo, then display \code{NA}.
}
\examples{
# sample of getting NHINo via RCFNo of Chang Gung Medical Hospital.
head(getNHINoViaRCFNo(df = sample_taiwan_drug, RCFNoColName = CGMH_CODE))
}
