% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getRxCuiInfo.R
\name{getRxCuiInfo}
\alias{getRxCuiInfo}
\title{Get RxCui information}
\usage{
getRxCuiInfo(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame, included RxCui}

\item{RxCuiColName}{a column name for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
This function can get the information of RxCui
}
