% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getRxcuiViaAtc.R
\name{getRxcuiViaAtc}
\alias{getRxcuiViaAtc}
\title{Get RxCui via ATC}
\usage{
getRxcuiViaAtc(df, AtcColName = ATC, cores = 8)
}
\arguments{
\item{df}{data.frame, a drug ingredient`s ATC of the specific indication}

\item{AtcColName}{a column name for ATC of df}

\item{cores}{number of parallel operation}

\item{Differ_ColName}{if colum for NHINo of df is not named ATC}
}
\value{
A new data.table \code{sample_atc_rxcui} with following columns.

\code{RxCui} drug RxCui

\code{ATC} drug ATC
}
\description{
\code{getRxcuiViaAtc} is a function to get RxCui via Anatomical Therapeutic Chemical Classification System(ATC)
}
\details{
This function provides user to get RxCui via ATC.
}
\examples{
# sample of getting RxCui via ATC.
sample_atc_rxcui <- getRxcuiViaAtc(df = sample_ATC,AtcColName = ATC,cores = 2)
head(sample_atc_rxcui)
}
