% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxcuiViaMayprevent.R
\name{getRxcuiViaMayprevent}
\alias{getRxcuiViaMayprevent}
\title{Get RxCui based on may prevent}
\usage{
getRxcuiViaMayprevent(strmayprevent)
}
\arguments{
\item{strmaytreat}{A may prevent
Details
This function provides user to get medicine ingredient RxCui via specific indication Mayprevent.}
}
\description{
Get RxCui based on may prevent
}
\examples{
# sample of searching an indication - esophagitis.
Esophagitis_List <- getRxcuiViaMayprevent("esophagitis")
# sample of getting RxCui list of esophagitis.
head(Esophagitis_List)
}
