% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getRxCuiViaMayTreat.R
\name{getRxCuiViaMayTreat}
\alias{getRxCuiViaMayTreat}
\title{Get ingredient RxCui via specific indication}
\usage{
getRxCuiViaMayTreat(strmaytreat)
}
\arguments{
\item{strmaytreat}{user can input an indication name}
}
\value{
A new data.table \code{maytreat_List} with following columns.

\code{RxCui} drug RxCui

\code{Name} drug name

\code{MinConcept.Id} indication Id

\code{MinConcept.Name} indication name
}
\description{
This function provides user to get drug ingredient RxCui of the specific indication may treat.
}
\details{
The data source is from the U.S. Veterans Administration's \code{MED-RT} database.
Since RxNorm integrates the US Veterans Administration MED-RT drug vocabulary and provides API (/rxclass/relaSources) that mediates RxCui into the drug vocabulary.
Therefore, we establishes a function through this API, so that users can obtain complete contraindications, indications and other information in the MED-RT vocabulary through the drug RxCui.
In addition to drug contraindications and indications for RxNorm coding, this study also uses web crawler technology to crawl information on contraindications and indications for all drugs in the MED-RT data sheet.
Establish a function for finding drugs from contraindications and indications.
}
\examples{
# sample of searching an indication esophagitis may treat.
Esophagitis_List <- getRxCuiViaMayTreat("esophagitis")
# sample of getting drug ingredient RxCui list of esophagitis.
head(Esophagitis_List)
}
