% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getRxcuiViaMaytreat.R
\name{getRxcuiViaMaytreat}
\alias{getRxcuiViaMaytreat}
\title{Get RxCui via may treat}
\usage{
getRxcuiViaMaytreat(strmaytreat)
}
\arguments{
\item{strmaytreat}{user can input an indication name.}
}
\value{
A new data.table \code{maytreat_List} with following columns.

\code{RxCui} drug RxCui

\code{Name} drug name

\code{MinConcept.Id} indication Id

\code{MinConcept.Name} indication name
}
\description{
\code{getRxcuiViaMaytreat} is a function to get RxCui via may treat.
}
\details{
The data source is from the U.S. Veterans Administration's \code{MED-RT} database. This function provides user to get drug ingredient`s RxCui of the specific indication may treat.
}
\examples{
# sample of searching an indication - esophagitis.
Esophagitis_List <- getRxcuiViaMaytreat("esophagitis")
# sample of getting drug RxCui list of esophagitis.
head(Esophagitis_List)
}
