% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRxcuiViaNdc.R
\name{getRxcuiViaNdc}
\alias{getRxcuiViaNdc}
\title{Get RxCui based on NDC}
\usage{
getRxcuiViaNdc(df, NdcColName = NDC, cores = 8)
}
\arguments{
\item{df}{data.frame include NDC}

\item{NdcColName}{A colum for NDC of df}

\item{cores}{number of parallel operation}
}
\description{
Get RxCui based on NDC
}
\details{
This function provides user to get RxCui via National Drug Code(NDC).
}
\examples{
#sample of getting RxCui via NDC.
sample_data_subset_rxcui <- getRxcuiViaNdc(df = sample_data_subset,NdcColName = NationalDrugCode)
head(sample_data_subset_rxcui)
}
