% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funPharm.R, R/getRxCuiViaNDC.R
\name{getRxCuiViaNDC}
\alias{getRxCuiViaNDC}
\title{Get RxCui via U.S National Drug Code(NDC)}
\usage{
getRxCuiViaNDC(df, NdcColName = NDC, cores = 8)
}
\arguments{
\item{df}{data.frame, a drug ingredient`s NDC of the specific indication}

\item{NdcColName}{a column name for NDC of df}

\item{cores}{number of parallel operation}
}
\value{
A new data.table \code{sample_data_subset_rxcui} with following columns.

\code{MemberId} patient Id

\code{DispenseDate} patient`s drug dispensed date, and the time interval is from 2025-12-06 to 2038-01-25

\code{NationalDrugCode} drug NDC, total of 17,326 different drug codes

\code{Quantity} quantity of patient`s drug dispensed

\code{DaysSupply} days of patient`s drug dispensed

\code{RxCui} drug code for the drug in the RxNorm drug vocabulary

\code{ndcStatus} NDC status
}
\description{
This is a function to get RxCui via National Drug Code(NDC)
}
\details{
User can input the data.frame with drug column which encoded in NDC.
The content is to add a column named RxCui in the original input data frame, storing the result of converting NDC to RxNorm in this column.
The column \code{ndcStatus} is to show whether the converted NDC is now used.
}
\examples{
#sample of getting RxCui via NDC.
sample_data_subset_rxcui <- getRxCuiViaNDC(df = sample_data_subset, NdcColName = NationalDrugCode)
head(sample_data_subset_rxcui)
}
