% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSbdRxcuiViaRxcui.R
\name{getSbdRxcuiViaRxcui}
\alias{getSbdRxcuiViaRxcui}
\title{Get Semantic Branded Drug RxCui  based on ingredient RxCui}
\usage{
getSbdRxcuiViaRxcui(df, RxCuiColName = RxCui, cores = 8)
}
\arguments{
\item{df}{data.frame include RxCui}

\item{RxCuiColName}{A colum for RxCui of df}

\item{cores}{number of parallel operation}
}
\description{
Get Semantic Branded Drug RxCui  based on ingredient RxCui
}
\details{
This function provides user to get Semantic Branded Drug(SBD) and Semantic Clinical Drug(SCD) RxCui via ingredient RxCui.
}
\examples{
# sample of getting Sbd Scd Rxcui via RxCui list of esophagitis.
Esophagitis_SBD_SCD_List <- getSbdRxcuiViaRxcui(df = Esophagitis_List[3,],RxCuiColName = RxCui,cores = 2)
head(Esophagitis_SBD_SCD_List)
}
